Do postprocessing in main rho process
This refactor has several benefits:
  - can share utility functions between postprocessing and the rest of rho
  - can use logging from postprocessing
  - will be able to split postprocessing into different modules in the future

The motiviation for doing it right now is that it also lets us do
multiple Ansible runs in one Rho scan, for instance to find the home
directory of an EAP installation and then run a playbook that depends
on it.

This particular change does not add or remove functionality; it just
moves around code that was already in rho.