Detect EAP_HOME directories and Fuse-on-EAP
Modify roles/jboss_eap/tasks/main.yml to detect EAP_HOME directories,
and add new tasks to look inside of the EAP_HOME directories for
indicators that Fuse-on-EAP is installed. Also add postprocessing for
the new results, and new fact names to trigger all of the above.

Note that we're starting to see some blowup in the when clauses in
that playbook. It's fixable, I just didn't want to do that fix in this
commit.